Add Dockerfile and update .goreleaser.yaml for multi-arch build
Introduced a Dockerfile for building the `autoproofcli` application.
This is done for creating an isolated environment for the application execution.

Also, updated the `.goreleaser.yaml` file to support multiple architectures i.e., amd64 and arm64.
Removed support for 386 architecture.

Docker images generation for each release is also added in the configuration.
This update is intended to extend the application's compatibility over different architectures and ease deployment process.