Refactor project configuration handling
This commit refactors the project configuration handling to separate
handling of public and secret options. The secret configuration is now
stored in a separate file located in the user's home directory which ensures
the security of sensitive data. Functions related to reading and modifying
the configuration are updated as well to properly deal with public and secret options.